backend/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
Pillow>=9.0.0
opencv-python>=4.5.0
pandas>=1.3.0
ultralytics>=8.0.0
pytest>=7.0.0
python-jose>=3.3.0
pydantic[email]>=1.8.0
mysqlclient>=2.0.0
passlib>=1.7.4
pydantic-settings>=2.0.0
bcrypt==3.2.0
SQLAlchemy>=1.4.0
websockets>=11.0.3